requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
transformers
torch
numpy
reportlab
deep-translator
pyvis
google-genai
beautifulsoup4